# swagger/output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Kitchen API",
    "description": "API бэкенда проектной кухни",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "http",
      "scheme": "bearer",
      "bearerFormat": "JWT"
    }
  },
  "paths": {
    "/api/user": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Получить данные пользователя [на фронте - getUser()]",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "headers",
            "description": "Token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "$ref": "#/definitions/Auth"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Обновить данные пользователя [на фронте - patchUser()]",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "headers",
            "description": "Token",
            "required": true,
            "type": "string"
          },
          {
            "name": "user",
            "in": "body",
            "description": "User data.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string",
                      "example": "John Smith"
                    },
                    "email": {
                      "type": "string",
                      "example": "john@gmail.com"
                    },
                    "bio": {
                      "type": "string",
                      "example": "My name is John"
                    },
                    "password": {
                      "type": "string",
                      "example": "newPassword123"
                    },
                    "image": {
                      "type": "string",
                      "example": "https://static.productionready.io/images/smiley-cyrus.jpg"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully updated.",
            "schema": {
              "$ref": "#/definitions/Auth"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/login": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Авторизоваться [на фронте - signIn()]",
        "description": "",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User email.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "example": "john@gmail.com"
                    },
                    "password": {
                      "type": "string",
                      "example": "myPassword"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully logged in.",
            "schema": {
              "$ref": "#/definitions/Auth"
            }
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/users": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Зарегистрироваться [на фронте - register()]",
        "description": "",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User data.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string",
                      "example": "John_Smith"
                    },
                    "email": {
                      "type": "string",
                      "example": "john@gmail.com"
                    },
                    "password": {
                      "type": "string",
                      "example": "password123"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully registered.",
            "schema": {
              "$ref": "#/definitions/Auth"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Auth": {
      "type": "object",
      "properties": {
        "user": {
          "type": "object",
          "properties": {
            "username": {
              "type": "string",
              "example": "John Smith"
            },
            "email": {
              "type": "string",
              "example": "john@gmail.com"
            },
            "token": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjYyNmVjZmNmNjQ4N2Y4MDAxNDJiNDFhYSIsInVzZXJuYW1lIjoiMTIzIiwiZXhwIjoxNjU2NjEzMzI3LCJpYXQiOjE2N"
            },
            "bio": {
              "type": "string",
              "example": "My name is John"
            },
            "image": {
              "type": "string",
              "example": "https://static.productionready.io/images/smiley-cyrus.jpg"
            }
          }
        }
      }
    },
...
  }
}